pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hero-images"
version = "0.1.0"
description = "Identifying aesthetically pleasing wildlife photos from camera trap datasets"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "Dan Morris"},
]
dependencies = [
    "pillow>=11.3.0",
    "google-generativeai>=0.8.5",
    "google-genai>=1.38.0",
    "requests>=2.31.0",
    "vllm>=0.3.0",
    "huggingface-hub>=0.19.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[project.scripts]
hero-gemini-batch = "hero_images.gemini_batch_labeling:main"
hero-ollama-local = "hero_images.ollama_local_labeling:main"
hero-vllm-local = "hero_images.vllm_local_labeling:main"
hero-generate-candidates = "hero_images.generate_sequence_aware_candidates_optimized:main"
hero-visualize = "hero_images.generate_label_visualization:main"

[tool.setuptools.packages.find]
include = ["hero_images*"]
